Fix: Move git installation to Brewfile

+++ install.sh
@@ -76,8 +76,8 @@ done;
 unset file;
 
 # install dotfiles via symbolic link
-for file in $HOME/dotfiles/dotfiles/*; do
-    if ask "Do you want to install $(basename "$file")?"; then
+for file in $HOME/dotfiles/dotfiles/.*; do
+    if test -e "$file" && ask "Do you want to install $(basename "$file")?"; then
         ln -s "$(realpath $file)" $HOME/$file;
     fi
 done;
